requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.1
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.4
Pillow==9.2.0
PyJWT==2.4.0
python-decouple==3.8
python-math==0.0.1
pytz==2022.2.1
requests==2.28.2
sqlparse==0.4.2
stripe==5.2.0
tzdata==2022.2
urllib3==1.26.14
